manifest.rsに移動
